Separate each experiment into its own file
